refactor: 우승자 출력 부분 새로운 클래스로 이동
